+++ app/models/concerns/grid_export.rb
@@ -3,10 +3,10 @@
 module GridExport
   extend ActiveSupport::Concern
 
-  def generate_csv_grids(sheet_scope, filename, raw_data, folder)
+  def generate_csv_grids(sheet_scope, temp_dir, filename, raw_data, folder)
     sheet_scope = sheet_scope.order(id: :desc)
-    tmp_export_file = File.join("tmp", "files", "exports", "#{filename}_grids_#{raw_data ? 'raw' : 'labeled'}_tmp.csv")
-    export_file = File.join("tmp", "files", "exports", "#{filename}_grids_#{raw_data ? 'raw' : 'labeled'}.csv")
+    tmp_export_file = File.join(temp_dir, "#{filename}_grids_#{raw_data ? "raw" : "labeled"}_tmp.csv")
+    export_file = File.join(temp_dir, "#{filename}_grids_#{raw_data ? "raw" : "labeled"}.csv")
     design_ids = sheet_scope.select(:design_id)
     grid_group_variables = all_design_variables_using_design_ids(design_ids).where(variable_type: "grid")
     sheet_ids = compute_sheet_ids_with_max_position(sheet_scope)
@@ -19,7 +19,7 @@ module GridExport
       load_all_grids(grid_group_variables, sheet_ids, raw_data, csv)
     end
     transpose_tmp_csv(tmp_export_file, export_file)
-    ["#{folder}/#{export_file.split('/').last}", export_file]
+    ["#{folder}/#{export_file.split("/").last}", export_file]
   end
 
   def grid_get_corresponding_names(sheet_ids, ids_and_names)